Refactor Navbar and CategoryFilter components
