+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/hasan/Downloads/microverse/bookstore/src/redux/serverUrl.js":"1","/home/hasan/Downloads/microverse/bookstore/src/index.js":"2","/home/hasan/Downloads/microverse/bookstore/src/components/Login.js":"3","/home/hasan/Downloads/microverse/bookstore/src/components/BookStoreHeader.js":"4","/home/hasan/Downloads/microverse/bookstore/src/container/BooksList.js":"5","/home/hasan/Downloads/microverse/bookstore/src/components/Loading.js":"6","/home/hasan/Downloads/microverse/bookstore/src/components/Book.js":"7","/home/hasan/Downloads/microverse/bookstore/src/redux/filter/filterActions.js":"8","/home/hasan/Downloads/microverse/bookstore/src/redux/booksFetch/booksActions.js":"9"},{"size":70,"mtime":1607086642420,"results":"10","hashOfConfig":"11"},{"size":811,"mtime":1607086772196,"results":"12","hashOfConfig":"11"},{"size":436,"mtime":1607076490638,"results":"13","hashOfConfig":"11"},{"size":618,"mtime":1606850127218,"results":"14","hashOfConfig":"11"},{"size":1049,"mtime":1607072191071,"results":"15","hashOfConfig":"11"},{"size":115,"mtime":1606941183805,"results":"16","hashOfConfig":"11"},{"size":2090,"mtime":1606988122041,"results":"17","hashOfConfig":"11"},{"size":150,"mtime":1606913948201,"results":"18","hashOfConfig":"11"},{"size":2185,"mtime":1607086622292,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"116wozx",{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":1,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/hasan/Downloads/microverse/bookstore/src/redux/serverUrl.js",[],"/home/hasan/Downloads/microverse/bookstore/src/index.js",[],"/home/hasan/Downloads/microverse/bookstore/src/components/Login.js",[],"/home/hasan/Downloads/microverse/bookstore/src/components/BookStoreHeader.js",[],"/home/hasan/Downloads/microverse/bookstore/src/container/BooksList.js",[],"/home/hasan/Downloads/microverse/bookstore/src/components/Loading.js",[],"/home/hasan/Downloads/microverse/bookstore/src/components/Book.js",[],"/home/hasan/Downloads/microverse/bookstore/src/redux/filter/filterActions.js",["38"],"/home/hasan/Downloads/microverse/bookstore/src/redux/booksFetch/booksActions.js",[],{"ruleId":"39","severity":2,"message":"40","line":3,"column":1,"nodeType":"41","endLine":8,"endColumn":4},"import/prefer-default-export","Prefer default export.","ExportNamedDeclaration"]
+[{"/home/hasan/Downloads/microverse/bookstore/src/components/CategoryFilter.js":"1"},{"size":525,"mtime":1607088651063,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"116wozx","/home/hasan/Downloads/microverse/bookstore/src/components/CategoryFilter.js",[]]
