+++ src/components/App.js
@@ -4,10 +4,8 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import BooksList from '../container/BooksList';
 import BooksForm from '../container/BooksForm';
-import BookStoreHeader from './BookStoreHeader';
 import { fetchUserAutoLogin } from '../redux';
-
-// import Login from './Login';
+import BookStoreHeader from './BookStoreHeader';
 
 function App() {
   const dispatch = useDispatch();
